add form

+++ src/app/components/person-list/person-list.component.ts
@@ -5,6 +5,8 @@ import { Person } from '../../shared/model/person';
 import { RouterLink } from '@angular/router';
 import { PersonService } from '../../services/person.service';
 import { MatButtonModule } from '@angular/material/button';
+import { MatDialog } from '@angular/material/dialog';
+import { DeletePersonDialogComponent } from '../delete-person-dialog/delete-person-dialog.component';
 
 @Component({
   selector: 'app-person-list',
@@ -13,20 +15,30 @@ import { MatButtonModule } from '@angular/material/button';
   templateUrl: './person-list.component.html',
   styleUrl: './person-list.component.css'
 })
-export class PersonListComponent implements OnInit {
+export class PersonListComponent implements OnInit{
 
-  constructor(private personService:PersonService){}
+
+ constructor(private personService: PersonService,private dialog: MatDialog) {}
+
   displayedColumns: string[] = ['fullName', 'email', 'actions'];
   persons: Person[]=[]
 
   ngOnInit(): void {
     this.persons=this.personService.list()
   }
-
-  deletePerson(id:number){
-    this.personService.delete(id)
-    this.persons=this.personService.list()
+  deletePerson(id:number,fullName:string){
+    const dialogRef = this.dialog.open(DeletePersonDialogComponent,{data: fullName,});
+    
+  dialogRef.afterClosed().subscribe(deletionConfirmed => {
+        if(deletionConfirmed==true){
+          this.personService.delete(id)
+          this.persons=this.personService.list()
+          }
+      });
   }
-
+ 
+ 
+ 
+  
 
 }
